requirements.txt
tensorflow==2.12.0
numpy
Pillow
streamlit
nltk
tqdm
